resolved POST request issue

+++ src/main/kotlin/org/openg2p/voucherservice/repository/VoucherProgramRepository.kt
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Component
 import org.springframework.stereotype.Repository
 
 @Repository
-interface VoucherProgramRepository : JpaRepository<VoucherProgram,Long>
+interface VoucherProgramRepository : JpaRepository<VoucherProgram,Int>
 {
 }
